updated install script

+++ install.sh
@@ -1,10 +1,9 @@
 #/bin/bash
-echo "backing up zshrc file" &&
-cp ~/.zshrc ~/.zshrc.copy &&
-echo "removing zshrc file" &&
-rm  -r ~/.zshrc &&
-echo "installing powerline fonts"
-./fonts/install.sh &&
-echo "installing new zhrc" &&
-ln -s  $(pwd)/zsh/zshconfig ~/.zshrc
-source ~/.zshrc
+if [ -f ~/.zshrc ]; then
+    echo "backing up zshrc file" &&
+    cp ~/.zshrc ~/.zshrc.copy &&
+    echo "removing zshrc file" &&
+    rm -r ~/.zshrc
+fi
+echo "installing new zshrc" &&
+ln -s  $(pwd)/zsh/zshrc ~/.zshrc
